better tests

+++ orders/resources.py
@@ -4,6 +4,7 @@ from datetime import datetime
 from flask import request
 from flask_restful import Resource, abort
 from sqlalchemy import desc
+from sqlalchemy.exc import SQLAlchemyError
 
 from . import db
 from .model import Worker, WorkOrder
@@ -11,19 +12,30 @@ from .schema import work_orders_schema, worker_schema
 from .validator import (validator_decorator, work_order_dic_schema,
                         worker_dic_schema)
 
+'''
+This is the hard of the RESTFul application.
+'''
+
 logger = logging.getLogger(__name__)
+MAX_NUMBER_OF_WORKERS_PER_WORK_ORDER = 5
 
 
 class WorkerResource(Resource):
     '''
-
+    All the worker related REST actions are coded here
     '''
 
     def get(self, worker_id):
         ''' not necessary but just for testing '''
-        worker = Worker.query.get(worker_id)
+        try:
+            worker = Worker.query.get(worker_id)
+        except SQLAlchemyError as e:
+            logger.error(e)
+            return 'Something wrong: {}'.format(str(e)), 500
         if not worker:
-            abort(404, message="Worker {} doesn't exist".format(worker_id))
+            message = "Worker {} does not exist".format(worker_id)
+            logger.warning(message)
+            abort(404, message=message)
         else:
             result = worker_schema.dump(worker)
             return result
@@ -39,22 +51,27 @@ class WorkerResource(Resource):
             )
             db.session.add(w)
             db.session.commit()
-            logger.debug("GET: {}".format(w))
+            logger.info("Worker created: {}".format(w))
             return w.id, 201
-        except Exception as e:
-            logger.error(str(e))
+        except SQLAlchemyError as e:
+            logger.error(e)
             return 'Something wrong: {}'.format(str(e)), 500
 
     def delete(self, worker_id):
         try:
             worker = Worker.query.get(worker_id)
-            logger.debug('Delete: {}'.format(worker))
-            db.session.delete(worker)
-            db.session.commit()
-            return '', 204
-        except Exception as e:
-            logger.error(str(e))
-            return 'Not Found {}'.format(worker_id), 404
+            if not worker:
+                message = "Worker {} does not exist".format(worker_id)
+                logger.warning(message)
+                abort(404, message=message)
+            else:
+                db.session.delete(worker)
+                db.session.commit()
+                logger.info('Worker deleted: {}'.format(worker))
+                return '', 204
+        except SQLAlchemyError as e:
+            logger.error(e)
+            return 'Something wrong: {}'.format(worker_id), 404
 
     def put(self, worker_id, work_order_id):
         ''' Assigning a worker to an order
@@ -64,17 +81,23 @@ class WorkerResource(Resource):
             n_workers = db.session.query(WorkOrder).join(
                 WorkOrder.workers).filter(WorkOrder.id ==
                                           work_order_id).count()
-            if n_workers >= 5:
-                return abort(404, message="Order is full")
+            if n_workers >= MAX_NUMBER_OF_WORKERS_PER_WORK_ORDER:
+                message = ("The work order {} is full. You only can have {} "
+                           "workers per work order".format(
+                               work_order_id,
+                               MAX_NUMBER_OF_WORKERS_PER_WORK_ORDER))
+                logger.warning(message)
+                return abort(404, message=message)
             else:
                 work_order = WorkOrder.query.get(work_order_id)
                 worker = Worker.query.get(worker_id)
                 worker.worker_orders.append(work_order)
                 db.session.commit()
-                return "Worker {} assigned to order {}.".format(
-                    worker, work_order), 200
-        except Exception as e:
-            logger.error(str(e))
+                logger.info("Worker {} assigned to order {}.".format(
+                    worker, work_order))
+                return '', 200
+        except SQLAlchemyError as e:
+            logger.error(e)
             return 'Something wrong happened: {}'.format(e), 404
 
 
@@ -84,20 +107,33 @@ class WorkOrderResource(Resource):
         ''' Fetch all work orders:
             - For a specific worker
             - Sorted by deadline '''
+        logger.debug("WorkOrderResource GET: {}".format(worker_id))
         if worker_id is None:
-            logger.debug("Getting all orders...")
-            orders = WorkOrder.query.order_by(desc(
-                WorkOrder.deadline)).all()
+            logger.debug("Getting all work orders...")
+            try:
+                orders = WorkOrder.query.order_by(desc(
+                    WorkOrder.deadline)).all()
+            except SQLAlchemyError as e:
+                logger.error(e)
+                return 'Something wrong happened: {}'.format(e), 404
+            result = work_orders_schema.dump(orders)
+            return result.data
+        else:
+            try:
+                orders = WorkOrder.query.filter(
+                    WorkOrder.workers.any(id=worker_id)).all()
+            except SQLAlchemyError as e:
+                orders = []
+                logger.error(e)
+                return 'Something wrong happened: {}'.format(e), 404
+            if not orders:
+                message = "No Work Order exists for worker {}".format(
+                    worker_id)
+                logger.warning(message)
+                abort(404, message=message)
             result = work_orders_schema.dump(orders)
             return result.data
 
-        else:
-            logger.debug("Getting order for worker: {}".format(worker_id))
-            orders = WorkOrder.query.filter(
-                WorkOrder.workers.any(id=worker_id)).all()
-            result = work_orders_schema.dump(orders)
-            # logger.debug(pformat(result.data))
-            return result.data
 
     @validator_decorator(work_order_dic_schema)
     def post(self):
@@ -112,8 +148,8 @@ class WorkOrderResource(Resource):
             )
             db.session.add(wo)
             db.session.commit()
-            logger.debug("POST: {}".format(wo))
+            logger.info("Work Orderer created: {}".format(wo))
             return wo.id, 201
-        except Exception as e:
-            logger.error(str(e))
+        except SQLAlchemyError as e:
+            logger.error(e)
             return 'Something wrong: {}'.format(str(e)), 500
